Fix cyclic dependency
Headers in ray/util should not depend on those in ray/common
