+++ src/ray/common/buffer.h
@@ -51,6 +51,8 @@ class Buffer {
   }
 };
 
+std::shared_ptr<Buffer> GenerateRandomBuffer();
+
 /// Represents a byte buffer in local memory.
 class LocalMemoryBuffer : public Buffer {
  public:
